Refactor tree components to eliminate duplicated code
Extract shared SVG rendering logic (configureSVG, renderTreeLinks,
renderTreeNodes, addSVGTextBackgrounds, setTooltipListeners) into
generic helpers in tree.ts. Unify EtymologyTooltip and
DescendantsTooltip into a single TreeTooltip component parameterized
by treeKind. Merge identical CSS files into Tree.css.

Net reduction of ~750 lines across the tree component files.
