--- client/src/ety/EtymologyTree.css
@@ -1,17 +1,0 @@
-.tree-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.tree {
-  margin: auto;
-  font-size: small;
-}
-
-.tree .lang {
-  font-size: smaller;
-}
-
-.tree .romanization {
-  font-size: smaller;
-}
