+++ client/src/ety/EtymologyTree.tsx
@@ -1,34 +1,27 @@
-import "./EtymologyTree.css";
+import "./Tree.css";
 import {
   Etymology,
   InterLangDescendants,
   Item,
   Lang,
-  term,
   TreeRequest,
 } from "../search/types";
 import { xMeanClusterLayout } from "./treeCluster";
-import EtymologyTooltip, {
-  setEtymologyTooltipListeners,
-} from "./EtymologyTooltip";
+import TreeTooltip from "./TreeTooltip";
 import { PositionKind, hideTooltip } from "./tooltip";
-import { BoundedHierarchyPointNode, langColor } from "./tree";
+import {
+  addSVGTextBackgrounds,
+  configureSVG,
+  renderTreeLinks,
+  renderTreeNodes,
+  setTooltipListeners,
+} from "./tree";
 import { TreeKind } from "../search/types";
 
-import { select, Selection } from "d3-selection";
-import { link, curveStepAfter } from "d3-shape";
-import {
-  hierarchy,
-  HierarchyPointLink,
-  HierarchyPointNode,
-} from "d3-hierarchy";
-import {
-  RefObject,
-  useRef,
-  useEffect,
-  MutableRefObject,
-  useState,
-} from "react";
+import { curveStepAfter } from "d3-shape";
+import { hierarchy, HierarchyPointNode } from "d3-hierarchy";
+import { select } from "d3-selection";
+import { useRef, useEffect, useState } from "react";
 
 interface EtymologyTreeProps {
   setSelectedLang: (lang: Lang | null) => void;
@@ -86,7 +79,6 @@ export default function EtymologyTree({
     );
 
     return () => {
-      // clear the previous svg
       select(svg).selectAll("*").remove();
       hideTooltip(tooltipRef, setShowTooltip);
       setShowTooltip(false);
@@ -109,7 +101,8 @@ export default function EtymologyTree({
   return (
     <div className="tree-container">
       <svg className="tree" ref={svgRef} />
-      <EtymologyTooltip
+      <TreeTooltip
+        treeKind={TreeKind.Etymology}
         setSelectedLang={setSelectedLang}
         setSelectedItem={setSelectedItem}
         selectedDescLangs={selectedDescLangs}
@@ -117,7 +110,7 @@ export default function EtymologyTree({
         setSelectedTreeKind={setSelectedTreeKind}
         showTooltip={showTooltip}
         setShowTooltip={setShowTooltip}
-        treeNode={tooltipTreeNode}
+        treeNode={tooltipTreeNode as HierarchyPointNode<Etymology | InterLangDescendants> | null}
         svgElement={tooltipSVGElement}
         positionKind={tooltipPositionKind}
         divRef={tooltipRef}
@@ -138,13 +131,11 @@ function etymologyTreeSVG(
   setTooltipTreeNode: (node: HierarchyPointNode<Etymology> | null) => void,
   setTooltipSVGElement: (svg: SVGElement | null) => void,
   setTooltipPositionKind: (positionKind: PositionKind) => void,
-  tooltipRef: RefObject<HTMLDivElement>,
-  tooltipShowTimeout: MutableRefObject<number | null>,
-  tooltipHideTimeout: MutableRefObject<number | null>
+  tooltipRef: React.RefObject<HTMLDivElement>,
+  tooltipShowTimeout: React.MutableRefObject<number | null>,
+  tooltipHideTimeout: React.MutableRefObject<number | null>
 ) {
-  // https://github.com/d3/d3-hierarchy#hierarchy
   const root = hierarchy<Etymology>(tree, (d: Etymology) => d.parents);
-
   root.sort((a, b) => b.data.etyOrder - a.data.etyOrder);
 
   const fontSize = svgElement
@@ -159,7 +150,6 @@ function etymologyTreeSVG(
 
   const pointRoot = layout(root);
 
-  // Center the tree horizontally.
   let x0 = Infinity;
   let x1 = -x0;
   pointRoot.each((d) => {
@@ -167,109 +157,25 @@ function etymologyTreeSVG(
     if (d.x < x0) x0 = d.x;
   });
 
-  // const nLeaves = root.leaves().length;
-  // const width = nLeaves * (dx + sep) - sep;
   const width = x1 - x0 + dx;
   const height = (root.height + 1) * dy;
-
   const viewBox = [x0 - dx / 2, -dy / 2, width, height];
 
-  // crispEdges implementation quality varies from browser to browser. It
-  // generally seems to work well but for example Windows Firefox renders
-  // random lines with 2px instead of 1px. Consider this as a solution:
-  // https://github.com/engray/subpixelFix.
-  const svg = select(svgElement)
-    .attr("version", "1.1")
-    .attr("xmlns", "http://www.w3.org/2000/svg")
-    .attr("xmlns:xlink", "http://www.w3.org/1999/xlink")
-    .attr("xmlns:xhtml", "http://www.w3.org/1999/xhtml")
-    .attr("viewBox", viewBox)
-    .attr("width", width)
-    .attr("height", height)
-    .attr(
-      "style",
-      `min-width: ${width}px; max-width: ${width}px; height: auto; height: intrinsic;`
-    )
-    .attr("shape-rendering", "crispEdges")
-    .attr("vector-effect", "non-scaling-stroke")
-    .attr("text-anchor", "middle")
-    .attr("text-rendering", "optimizeLegibility")
-    // this noop event listener is to cajole mobile browsers (or, at least,
-    // ios webkit) into responding to touch events on svg elements, cf.
-    // https://stackoverflow.com/a/65777666/10658294
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    .on("touchstart", () => {});
+  const svg = configureSVG(svgElement, viewBox, width, height);
 
-  // the lines forming the tree
-  svg
-    .append("g")
-    .attr("fill", "none")
-    .attr("stroke", "#555")
-    .attr("stroke-opacity", 1.0)
-    .attr("stroke-linecap", "butt")
-    .attr("stroke-linejoin", "miter")
-    .attr("stroke-width", 1.0)
-    .selectAll("path")
-    .data(pointRoot.links())
-    .join("path")
-    .attr(
-      "d",
-      link<HierarchyPointLink<Etymology>, HierarchyPointNode<Etymology>>(
-        curveStepAfter
-      )
-        .x((d) => d.x)
-        .y((d) => d.y)
-    );
+  renderTreeLinks(svg, pointRoot, curveStepAfter, {
+    x: (d) => d.x,
+    y: (d) => d.y,
+  });
 
-  // Confusingly, with respect to the tree structure and d3 api, these are
-  // descendants. But with respect to the etymology, they are ancestors.
-  const ancestors: BoundedHierarchyPointNode<Etymology>[] = pointRoot
-    .descendants()
-    .map(function (d) {
-      return { node: d, bbox: new DOMRect(0, 0, 0, 0) };
-    });
+  const { node, nodeBackground } = renderTreeNodes(
+    svg,
+    pointRoot,
+    treeRootItem,
+    (n) => [n.x, n.y]
+  );
 
-  // placeholder rects for text backgrounds to be set in addSVGTextBackgrounds()
-  const nodeBackground = svg
-    .append("g")
-    .selectAll<SVGRectElement, unknown>("rect")
-    .data(ancestors)
-    .join("rect")
-    .attr("fill", "white");
-
-  // the text nodes
-  const node = svg
-    .append("g")
-    .selectAll<SVGTextElement, unknown>("g")
-    .data(ancestors)
-    .join("g")
-    .attr("font-weight", (d) =>
-      d.node.data.item.id === treeRootItem.id ? "bold" : null
-    )
-    .attr("transform", (d) => `translate(${d.node.x},${d.node.y})`);
-
-  node
-    .append("text")
-    .attr("class", "lang")
-    .attr("y", "-1em")
-    .attr("fill", (d) => langColor(d.node.data.langDistance))
-    .text((d) => d.node.data.item.lang.name);
-
-  node
-    .append("text")
-    .attr("class", "term")
-    .attr("y", "0.25em")
-    .text((d) => term(d.node.data.item));
-
-  node
-    .append("text")
-    .attr("class", "romanization")
-    .attr("y", "1.5em")
-    .text((d) =>
-      d.node.data.item.romanization ? `(${d.node.data.item.romanization})` : ""
-    );
-
-  setEtymologyTooltipListeners(
+  setTooltipListeners(
     node,
     setShowTooltip,
     setTooltipTreeNode,
@@ -280,38 +186,5 @@ function etymologyTreeSVG(
     tooltipHideTimeout
   );
 
-  addSVGTextBackgrounds(node, nodeBackground);
-}
-
-function addSVGTextBackgrounds(
-  node: Selection<
-    SVGGElement | SVGTextElement,
-    BoundedHierarchyPointNode<Etymology>,
-    SVGGElement,
-    undefined
-  >,
-  nodeBackground: Selection<
-    SVGRectElement,
-    BoundedHierarchyPointNode<Etymology>,
-    SVGGElement,
-    undefined
-  >
-) {
-  node.each(function (d) {
-    d.bbox = this.getBBox();
-  });
-
-  const xMargin = 3;
-  const yMargin = 3;
-
-  nodeBackground
-    .attr("width", (d) => d.bbox.width + 2 * xMargin)
-    .attr("height", (d) => d.bbox.height + 2 * yMargin)
-    .attr("transform", function (d) {
-      const x = d.node.x - xMargin;
-      const y = d.node.y - yMargin;
-      return `translate(${x},${y})`;
-    })
-    .attr("x", (d) => d.bbox.x)
-    .attr("y", (d) => d.bbox.y);
+  addSVGTextBackgrounds(node, nodeBackground, (d) => [d.node.x, d.node.y]);
 }
